Convert CRA to next.js app
