+++ frontend/components/LandingPage/Header.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface HeaderProps {}
+
+function Header(props: HeaderProps) {
+    return <h1 className="text-4xl font-bold">Welcome to PeerPrep</h1>;
+}
+
+export default Header;
