--- src/components/LandingPage/DescriptionModal/descriptionmodal.style.js
@@ -1,28 +1,0 @@
-const styles = {
-    modal: {
-        display: 'none',
-        position: 'fixed',
-        zIndex: 1,
-        left: '0',
-        top: '0',
-        width: '100%',
-        height: '100%',
-        overflow: 'auto',
-        backgroundColor: 'rgba(0,0,0,0.4)'
-    },
-    modalContent: {
-        backgroundColor: '#fefefe',
-        margin: '15% auto',
-        padding: '20px',
-        border: '1px solid #888',
-        width: '80%'
-    },
-    closeBtn: {
-        color: '#aaa',
-        float: 'right',
-        fontSize: '28px',
-        fontWeight: 'bold',
-    }
-};
-
-export default styles;
