--- src/components/LandingPage/Header/Header.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Header() {
-    return <h1>Welcome to PeerPrep</h1>;
-}
-
-export default Header;
