--- src/components/LandingPage/QuestionForm/questionform.styles.js
@@ -1,9 +1,0 @@
-const styles = {
-    textarea: {
-        whiteSpace: 'pre-wrap',
-        wordWrap: 'break-word', 
-        resize: 'vertical'
-    }
-};
-
-export default styles;
