--- src/components/LandingPage/QuestionTable/questiontable.styles.js
@@ -1,28 +1,0 @@
-const styles = {
-    table: {
-        borderCollapse: 'collapse',
-        width: '100%',
-        border: '3px solid black',
-        marginTop: '20px',
-        marginBottom: '20px'
-    },
-    th: {
-        backgroundColor: '#f2f2f2',
-        textAlign: 'left',
-        padding: '12px',
-        border: '1px solid black'
-    },
-    td: {
-        textAlign: 'left',
-        padding: '12px',
-        border: '1px solid black'
-    },
-    truncate: {
-        whiteSpace: 'nowrap',
-        overflow: 'hidden',
-        textOverflow: 'ellipsis',
-        maxWidth: '150px'
-    }
-};
-
-export default styles;
